push push

+++ database/migrations/2022_06_11_020847_create_buku_kas_table.php
@@ -18,7 +18,7 @@ return new class extends Migration
 
             $table->string('uraian', 191)->nullable();
             $table->integer('harga')->nullable();
-            $table->integer('volume')->nullable();
+            $table->float('volume')->nullable();
             $table->string('satuan', 191)->nullable();
             $table->string('noBukti', 191);
             $table->integer('penerimaan')->nullable();
